prototype/backend/requirements.txt
uvicorn
openpyxl
fastapi
langgraph
langchain 
langchain_openai
python-dotenv